refactor: Nest teams tools.
