+++ src/tools/index.ts
@@ -16,8 +16,6 @@ import "./financial-commitment-reports";
 import "./folders";
 import "./get-cost-provider-accounts";
 import "./get-financial-commitment-report";
-import "./get-team";
-import "./get-teams";
 import "./list-cost-integrations";
 import "./list-cost-providers";
 import "./list-cost-services";
@@ -30,6 +28,7 @@ import "./report-notifications";
 import "./resource-reports";
 import "./scenario-models";
 import "./tags";
+import "./teams";
 import "./user-feedback";
 import "./users";
 import "./workspaces";
