+++ src/tools/teams/index.ts
@@ -1,0 +1,2 @@
+import "./get-team";
+import "./get-teams";
